+++ test/unit/coreHandlers/handleScope-unit.test.ts
@@ -1,7 +1,7 @@
 import { getCurrentHub } from '@sentry/core';
 import { mockSdk } from '@test';
 
-import * as HandleScope from './handleScope';
+import * as HandleScope from '../../../src/coreHandlers/handleScope';
 
 let mockHandleScope: jest.MockedFunction<typeof HandleScope.handleScope>;
