+++ test/unit/eventBuffer.test.ts
@@ -3,7 +3,10 @@ import 'jsdom-worker';
 import { BASE_TIMESTAMP } from '@test';
 import pako from 'pako';
 
-import { createEventBuffer, EventBufferCompressionWorker } from './eventBuffer';
+import {
+  createEventBuffer,
+  EventBufferCompressionWorker,
+} from './../../src/eventBuffer';
 
 const TEST_EVENT = { data: {}, timestamp: BASE_TIMESTAMP, type: 3 };
