+++ test/unit/flush.test.ts
@@ -1,10 +1,10 @@
 import * as SentryUtils from '@sentry/utils';
 import { BASE_TIMESTAMP, mockRrweb, mockSdk } from '@test';
 
-import { useFakeTimers } from './../test/utils/use-fake-timers';
-import { SESSION_IDLE_DURATION } from './session/constants';
-import { createPerformanceEntries } from './createPerformanceEntry';
-import { Replay } from './';
+import { Replay } from './../../src';
+import { createPerformanceEntries } from './../../src/createPerformanceEntry';
+import { SESSION_IDLE_DURATION } from './../../src/session/constants';
+import { useFakeTimers } from './../../test/utils/use-fake-timers';
 
 useFakeTimers();
