+++ test/unit/session/deleteSession.test.ts
@@ -1,5 +1,5 @@
-import { REPLAY_SESSION_KEY } from './constants';
-import { deleteSession } from './deleteSession';
+import { REPLAY_SESSION_KEY } from '../../../src/session/constants';
+import { deleteSession } from '../../../src/session/deleteSession';
 
 const storageEngine = window.sessionStorage;
