+++ test/unit/util/isSampled.test.ts
@@ -1,4 +1,4 @@
-import { isSampled } from './isSampled';
+import { isSampled } from '../../../src/util/isSampled';
 
 // Note Math.random generates a value from 0 (inclusive) to <1 (1 exclusive).
 const cases = [
